# specs/openapi/master-data.v1.yaml
openapi: 3.1.0
info:
  title: HIMS Master Data API
  version: 1.0.0
  description: 'HIMS Master Data catalog API — platform reference data (modules first; more surfaces phased in). Base path
    `/api/v1/master-data`. DELETE on modules is a recursive soft-delete (`is_deleted = true`). Gateway-level authentication
    (API keys / JWT) is expected before production; this spec does not model per-route security yet.

    '
servers:
- url: /api/v1/master-data
paths:
...
  /visitpad/vitals:
    get:
      tags:
      - Visitpad — Vitals
      summary: List vitals
      operationId: get_vitals_api_v1_master_data_visitpad_vitals_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      - name: category
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/VisitpadVitalCategory'
          - type: 'null'
          title: Category
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitpadVitalListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Visitpad — Vitals
      summary: Create vital
      operationId: post_vital_api_v1_master_data_visitpad_vitals_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitpadVitalCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitpadVitalSingleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /visitpad/vitals/{vital_id}:
    get:
      tags:
      - Visitpad — Vitals
      summary: Get vital
      operationId: get_vital_api_v1_master_data_visitpad_vitals__vital_id__get
      parameters:
      - name: vital_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Vital Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitpadVitalSingleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Visitpad — Vitals
      summary: Update vital
      operationId: patch_vital_api_v1_master_data_visitpad_vitals__vital_id__patch
      parameters:
      - name: vital_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Vital Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitpadVitalUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitpadVitalSingleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Visitpad — Vitals
      summary: Soft-delete vital
      operationId: delete_vital_api_v1_master_data_visitpad_vitals__vital_id__delete
      parameters:
      - name: vital_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Vital Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitpadVitalSingleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    VisitpadVitalCategory:
      type: string
      enum:
      - vital_signs
      - anthropometric
      - functional
      - score
      - other
      title: VisitpadVitalCategory
    VisitpadVitalCreate:
      properties:
        code:
          type: string
          maxLength: 64
          minLength: 1
          title: Code
        name:
          type: string
          maxLength: 256
          minLength: 1
          title: Name
        short_name:
          type: string
          maxLength: 64
          minLength: 1
          title: Short Name
        category:
          $ref: '#/components/schemas/VisitpadVitalCategory'
        data_type:
          $ref: '#/components/schemas/VisitpadVitalDataType'
        unit:
          type: string
          maxLength: 128
          minLength: 1
          title: Unit
        default_unit_code:
          type: string
          maxLength: 64
          minLength: 1
          title: Default Unit Code
        allowed_units:
          items:
            type: string
          type: array
          maxItems: 32
          title: Allowed Units
        critical_low:
          anyOf:
          - type: number
          - type: 'null'
          title: Critical Low
        critical_high:
          anyOf:
          - type: number
          - type: 'null'
          title: Critical High
        reference_kind:
          $ref: '#/components/schemas/VisitpadVitalReferenceKind'
        reference_json:
          additionalProperties: true
          type: object
          title: Reference Json
        normal_range_adult:
          additionalProperties: true
          type: object
          title: Normal Range Adult
        normal_range_paediatric:
          additionalProperties: true
          type: object
          title: Normal Range Paediatric
        input_method:
          $ref: '#/components/schemas/VisitpadVitalInputMethod'
        is_paired:
          type: boolean
          title: Is Paired
          default: false
        pair_code:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Pair Code
        display_order:
          type: integer
          title: Display Order
          default: 0
        is_active:
          type: boolean
          title: Is Active
          default: true
        loinc_code:
          anyOf:
          - type: string
            maxLength: 32
          - type: 'null'
          title: Loinc Code
        snomed_observable_code:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Snomed Observable Code
      additionalProperties: false
      type: object
      required:
      - code
      - name
      - short_name
      - category
      - data_type
      - unit
      - default_unit_code
      - reference_kind
      - input_method
      title: VisitpadVitalCreate
    VisitpadVitalDataType:
      type: string
      enum:
      - numeric
      - text
      - boolean
      - score
      title: VisitpadVitalDataType
    VisitpadVitalInputMethod:
      type: string
      enum:
      - manual
      - device
      - calculated
      title: VisitpadVitalInputMethod
    VisitpadVitalListResponse:
      properties:
        data:
          items:
            $ref: '#/components/schemas/VisitpadVitalResponse'
          type: array
          title: Data
        total:
          type: integer
          title: Total
      type: object
      required:
      - data
      - total
      title: VisitpadVitalListResponse
    VisitpadVitalReferenceKind:
      type: string
      enum:
      - none
      - range
      - categorical
      - boolean
      title: VisitpadVitalReferenceKind
    VisitpadVitalResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        tenant_id:
          type: string
          format: uuid
          title: Tenant Id
        code:
          type: string
          title: Code
        name:
          type: string
          title: Name
        short_name:
          type: string
          title: Short Name
        category:
          $ref: '#/components/schemas/VisitpadVitalCategory'
        data_type:
          $ref: '#/components/schemas/VisitpadVitalDataType'
        unit:
          type: string
          title: Unit
        default_unit_code:
          type: string
          title: Default Unit Code
        allowed_units:
          items: {}
          type: array
          title: Allowed Units
        critical_low:
          anyOf:
          - type: number
          - type: 'null'
          title: Critical Low
        critical_high:
          anyOf:
          - type: number
          - type: 'null'
          title: Critical High
        reference_kind:
          $ref: '#/components/schemas/VisitpadVitalReferenceKind'
        reference_json:
          additionalProperties: true
          type: object
          title: Reference Json
        normal_range_adult:
          additionalProperties: true
          type: object
          title: Normal Range Adult
        normal_range_paediatric:
          additionalProperties: true
          type: object
          title: Normal Range Paediatric
        input_method:
          $ref: '#/components/schemas/VisitpadVitalInputMethod'
        is_paired:
          type: boolean
          title: Is Paired
        pair_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Pair Code
        display_order:
          type: integer
          title: Display Order
        is_active:
          type: boolean
          title: Is Active
        is_deleted:
          type: boolean
          title: Is Deleted
        loinc_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Loinc Code
        snomed_observable_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Snomed Observable Code
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - tenant_id
      - code
      - name
      - short_name
      - category
      - data_type
      - unit
      - default_unit_code
      - allowed_units
      - reference_kind
      - reference_json
      - normal_range_adult
      - normal_range_paediatric
      - input_method
      - is_paired
      - display_order
      - is_active
      - is_deleted
      - created_at
      - updated_at
      title: VisitpadVitalResponse
    VisitpadVitalSingleResponse:
      properties:
        data:
          $ref: '#/components/schemas/VisitpadVitalResponse'
      type: object
      required:
      - data
      title: VisitpadVitalSingleResponse
    VisitpadVitalUpdate:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 256
            minLength: 1
          - type: 'null'
          title: Name
        short_name:
          anyOf:
          - type: string
            maxLength: 64
            minLength: 1
          - type: 'null'
          title: Short Name
        category:
          anyOf:
          - $ref: '#/components/schemas/VisitpadVitalCategory'
          - type: 'null'
        data_type:
          anyOf:
          - $ref: '#/components/schemas/VisitpadVitalDataType'
          - type: 'null'
        unit:
          anyOf:
          - type: string
            maxLength: 128
            minLength: 1
          - type: 'null'
          title: Unit
        default_unit_code:
          anyOf:
          - type: string
            maxLength: 64
            minLength: 1
          - type: 'null'
          title: Default Unit Code
        allowed_units:
          anyOf:
          - items:
              type: string
            type: array
            maxItems: 32
          - type: 'null'
          title: Allowed Units
        critical_low:
          anyOf:
          - type: number
          - type: 'null'
          title: Critical Low
        critical_high:
          anyOf:
          - type: number
          - type: 'null'
          title: Critical High
        reference_kind:
          anyOf:
          - $ref: '#/components/schemas/VisitpadVitalReferenceKind'
          - type: 'null'
        reference_json:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Reference Json
        normal_range_adult:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Normal Range Adult
        normal_range_paediatric:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Normal Range Paediatric
        input_method:
          anyOf:
          - $ref: '#/components/schemas/VisitpadVitalInputMethod'
          - type: 'null'
        is_paired:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Paired
        pair_code:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Pair Code
        display_order:
          anyOf:
          - type: integer
          - type: 'null'
          title: Display Order
        is_active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
        loinc_code:
          anyOf:
          - type: string
            maxLength: 32
          - type: 'null'
          title: Loinc Code
        snomed_observable_code:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Snomed Observable Code
        is_deleted:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Deleted
      additionalProperties: false
      type: object
      title: VisitpadVitalUpdate